Respect app config when validate phone number
