+++ pkg/auth/handler/webapp/authflowv2/login.go
@@ -176,7 +176,7 @@ func (h *AuthflowV2LoginHandler) ServeHTTP(w http.ResponseWriter, r *http.Reques
 
 		loginID := r.Form.Get("x_login_id")
 		loginIDInputType := r.Form.Get("x_login_id_input_type")
-		identification := webapp.GetMostAppropriateIdentification(screen.StateTokenFlowResponse, loginID, loginIDInputType)
+		identification := webapp.GetMostAppropriateIdentification(ctx, screen.StateTokenFlowResponse, loginID, loginIDInputType)
 		input := map[string]interface{}{
 			"identification": identification,
 			"login_id":       loginID,
